requirements.txt
newsapi-python==0.2.6
python-decouple==3.6
python-telegram-bot==20.0a0